refactor(list): encapsulate usize with u64

+++ src/list.rs
@@ -29,7 +29,7 @@ pub fn should_list_file(path: &Path, cli: &Cli, extensions: &Option<Vec<String>>
 
     if let Some(max_size) = cli.max_size {
         if let Ok(metadata) = path.metadata() {
-            if metadata.len() > max_size * 1024 * 1024 {
+            if metadata.len() > (max_size as u64) * 1024  * 1024 {
                 return false;
             }
         }
